move map to collection
