+++ java/fury-core/src/main/java/io/fury/resolver/ClassResolver.java
@@ -76,8 +76,8 @@ import io.fury.serializer.UnexistedClassSerializers.UnexistedSkipClass;
 import io.fury.serializer.UnmodifiableSerializers;
 import io.fury.serializer.collection.CollectionSerializer;
 import io.fury.serializer.collection.CollectionSerializers;
-import io.fury.serializer.map.MapSerializer;
-import io.fury.serializer.map.MapSerializers;
+import io.fury.serializer.collection.MapSerializer;
+import io.fury.serializer.collection.MapSerializers;
 import io.fury.serializer.scala.SingletonObjectSerializer;
 import io.fury.type.ClassDef;
 import io.fury.type.Descriptor;
