+++ java/fury-core/src/main/java/io/fury/serializer/ImmutableCollectionSerializers.java
@@ -19,7 +19,7 @@ package io.fury.serializer;
 import io.fury.Fury;
 import io.fury.memory.MemoryBuffer;
 import io.fury.serializer.collection.CollectionSerializer;
-import io.fury.serializer.map.MapSerializer;
+import io.fury.serializer.collection.MapSerializer;
 import io.fury.util.Platform;
 import io.fury.util.unsafe._JDKAccess;
 import java.lang.invoke.MethodHandle;
