+++ java/fury-core/src/test/java/io/fury/serializer/map/MapSerializersTest.java
@@ -27,6 +27,7 @@ import io.fury.FuryTestBase;
 import io.fury.collection.MapEntry;
 import io.fury.config.Language;
 import io.fury.serializer.collection.CollectionSerializersTest.TestEnum;
+import io.fury.serializer.collection.MapSerializers;
 import io.fury.test.bean.MapFields;
 import io.fury.type.GenericType;
 import java.io.Serializable;
